Refactor the code after Nikolche code review.
